refactor(define-options): disallow more keys
