+++ .changeset/thirty-kids-rule.md
@@ -1,0 +1,5 @@
+---
+'unplugin-vue-define-options': patch
+---
+
+disallow more keys for `defineOptions`
