+++ packages/define-options/src/core/utils.ts
@@ -21,6 +21,9 @@ export function hasPropsOrEmits(node: ObjectExpression) {
     (prop) =>
       (prop.type === 'ObjectProperty' || prop.type === 'ObjectMethod') &&
       prop.key.type === 'Identifier' &&
-      (prop.key.name === 'props' || prop.key.name === 'emits')
+      (prop.key.name === 'props' ||
+        prop.key.name === 'emits' ||
+        prop.key.name === 'expose' ||
+        prop.key.name === 'slots')
   )
 }
