+++ packages/define-options/src/index.ts
@@ -8,7 +8,7 @@ import {
   createFilter,
   detectVueVersion,
 } from '@vue-macros/common'
-import { transformDefineOptions } from './core/transform'
+import { transformDefineOptions } from './core'
 
 export type Options = BaseOptions
 export type OptionsResolved = MarkRequired<Options, 'include' | 'version'>
